Move qualcomm.passes to qualcomm._passes (#5920)
Summary:
Changing qualcomm.passes to qualcomm._passes to indicate that these passes are not covered under the API stability guarantee.

Pull Request resolved: https://github.com/pytorch/executorch/pull/5920

Reviewed By: helunwencser

Differential Revision: D63926845

fbshipit-source-id: f1e3187f3012f83b388d9635ca5a244fa36950d6
(cherry picked from commit 866b40c72e5e8814e55dbb253e356366db282267)
