+++ backends/qualcomm/_passes/annotate_and_quant_scalar.py
