+++ backends/qualcomm/_passes/annotate_quant_attrs.py
