+++ backends/qualcomm/_passes/convert_binary_op_with_scalar.py
