+++ backends/qualcomm/_passes/utils.py
